pyproject.toml
[project]
name = "wordle-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
]
